+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - TasMesh support for LWT messages (#20392)
 - Show calculated heat index if temperature and humidity is available with ``#define USE_HEAT_INDEX`` (#4771)
 - Berry add explicit error log when memory allocation fails (#20807)
+- Support for AMS5915/AMS6915 temperature and pressure sensors (#20814)
 
 ### Breaking Changed
 
@@ -27,7 +28,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - ESP32 PWM activity on unconfigured PWM GPIOs (#20732)
 - Shutter inverted using internal commands (#20752)
-- HASPmota PSRAM memory leak
+- HASPmota PSRAM memory leak (#20818)
 
 ### Removed